requirements.txt
absl-py==2.3.1
attrs==25.4.0
cffi==2.0.0
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.61.0
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.9
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.4
narwhals==2.13.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==12.0.0
plotly==6.5.0
protobuf==4.25.8
pycparser==2.23
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pyVHR==2.0
scipy==1.15.3
sentencepiece==0.2.1
six==1.17.0
sounddevice==0.5.3
